[hid-input-report] Use LLCPP BufferAllocator
Use the new LLCPP BufferAllocator to allocate
DeviceDescriptors and InputReports.

The hid-input-report driver now stores one
large buffer which it allocates from to
create InputReports. This buffer is
cleared when the reports are sent to
a client.

The hid-input-report driver's device
descriptor is now stack allocated.

The old functions are going to be removed
in a later commit.

TEST: lib-hid-input-report-test
      hid-input-report-test
      Manually tested on hardware.

Bug: 49628
Change-Id: I5b45af307e6be5acf6cdb2dacb2f92c9891b8f6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382273
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>