[Harvester] Reduce warning messages to verbose level
The Harvester now caches a list of jobs/processes/threads to amortize the high
cost of gathering the list. This means that items in the list can be out of date
(increasing as the cache ages) and lead to more spamy log messages about
failing to get task information. This CL changes the log level of those messages
from WARNING to VERBOSE3. So the messages can still be seen if verbose3 logging
is requested.

bug=48250

Change-Id: Ic5c3c467693ddaa724aad379ac1f5cee3c746375
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386253
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Jayson Adams <shrike@google.com>
Testability-Review: Casey Dahlin <sadmac@google.com>