[dhcpv6] Define FIDL API for the client
Acquiring DNS addresses in stateless DHCP is the first use case we want
to deliver, and this API enables that.

`ClientProvider` is kept minimal. This way if in the future we want to
run one client per component, the change needed to `Client` protocol is
also minimal.

In the future we'll extend this API in 2 ways:
1. Allow starting DHCPv6 client in other models
2. Compose other watchers, e.g. an AddressWatcher

Fixed: 48861

Change-Id: I5353624f75be94318417e1bd269e9916446dd70d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387737
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Ghanan Gowripalan <ghanan@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jay Zhuang <jayzhuang@google.com>