[test_spec] `target` should be a target label
Previously this was the name of a local label (e.g. "foo"), but actually a label (e.g. ":foo").
For one thing this didn't allow for test_spec to specify a target anywhere else in the tree other than the same directory that the test_spec itself is defined.

Change-Id: I869a467bdfcb140233bff88474cbebf87ce66528
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388231
Reviewed-by: Gary Miguel <garymm@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Shai Barack <shayba@google.com>