[intl] Examine the assets directory
Adds the code that examines the assets directory and returns
the locales found there.  This is one of a few things needed
to provide asset loading on demand.

This was made to be a self-contained change for ease of
review.

Bug: 50725
Change-Id: Ie7f5c9f5e7eafc5cfb362f9f1a1752385c1b2712
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387476
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Testability-Review: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>