[intl] Factors out Model from json
Should be a pure refactoring.

json crate used to be intertwined with parser, which in turn is using
host-only libraries.  And rightfully so, since we only ever want to
prepare host tools on the host.

However, the loading of JSON libraries and their interpretation is also
the task of the code on the target.

This commit factors out the respective code and makes some modifications
that make this change possible.  `Model::from_dictionaries` had to
become a stand-alone function in `json`, because it needs host-only
`parser::Dictionary`.

Bug: 50725
Change-Id: Ie8973300ef50949ec187dd41ed2414ce9e4bd3ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387525
Testability-Review: Filip Filmar <fmil@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>