[fs] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow PseudoDir::Entry objects to exist in multiple WAVL trees
simultaneously.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Change-Id: I5f5f3cc29d1dcda90b957e23ea1bb03caa297683
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386820
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>