[kernel][vdso] Sanity-check the vDSO image matches the kernel
Extract the build ID note contents at build time and embed that
separately in the kernel so it can verify at runtime that it
matches the vDSO image.  The kernel bakes in expectations about
the exact instruction sequences that appear in the vDSO, so it's
crucial that it always be the precise vDSO image that the kernel
was built for.  The image is pasted into the kernel binary at
build time today, but in the future it will be packed differently
where accidental mismatches will become possible if the wrong ZBI
repacking shenanigans are done.

Test: change consists of adding an assert on the boot path
Change-Id: I61cdc42302fd11f09f7ffd2aa4e0a1ac04369a81
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383644
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>