[debugger] Skip compiler generated code.
If a backtrace or stop position is on an instruction marked in the line
table as compiler-generated, previously we would show "no line info" in
the stacks and assembly in the stop information.

This changes the Address->Line querying function to return the next line
in the line table when this happens.

This does not change the stepping logic because the line-table lookup
for that is a completely separate code path that calls into LLVM
directly. That code already handled these "line 0" cases properly.

Change-Id: I7d9bc89e8cbbf4b2acce525f84d8973b176efbb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389337
Testability-Review: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>