[fidl][rust] Report epitaph on write failure
Due to the shape of the API, epitaphs can only be reported
when a FIDL method that expects a response is sent.

Fire-and-forget FIDL methods are not `async` so there is no
way to wait for receiving an epitaph. An API change is needed
to support this.

Change-Id: I728e75f8efc20a5afc571f2c6867b76a78fe2b74
Bug: 50730
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386377
Reviewed-by: Marc Khouri <mnck@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>