[unification] Add fidl_library_json_rspfile()
This CL prepares future modifications that will implement
the ZN link_output_rspfile() template to the GN build.

This CL replaces the use of link_output_rspfile() in
//zircon/tools/vdso:json.rspfile with a new template named
fidl_library_json_rspfile(), used to perform the same thing,
i.e. generate a response file that contains the path to a
FIDL IR JSON file.

The reasons for this are:

  - link_output_rspfile() is documented as being used
    for linkable targets (e.g. executables, shared libraries
    and loadable modules), and an fidl.json file certainly
    doesn't look like one, so this ends up less confusing.

  - The GN and SDK builds' implementation of fidl_library()
    already generate the corresponding metadata with key
    "fidl_json", so using it here too seems logical, and
    less surprising overall.

    This will simplify migrating //zircon/tools/vdso to the
    GN build in the future.

  - This makes it slightly easier to implement
    link_output_rspfile() in the context of the GN build,
    which handles linkable targets very differently from
    the ZN one.

For testing, I manually verified that the content of
out/default.zircon/gen/vdso/json.rspfile was identical
before and after the CL and a clean build. For the record
it's a single line of text with value:

  gen/vdso/zx.json/fidl.json

Bug: 51971
Testing: Build only change, see above for manual test procedure.
Change-Id: I5b3f69ba02f73108bc0c9f94d3d4ede9fbb2cfe6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389233
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>