[testsharder] Add host-test paths to shard deps
This replaces identical logic in recipes, specifically lines 805-807
here:
https://fuchsia.googlesource.com/infra/recipes/+/91b211a1157662f261ab90e287a13fc36483386e/recipe_modules/testing_requests/api.py#805

Note that as long as that for loop is still present, the recipes will
call `register_link` twice on each host test path: once in the
`shard.tests` loop and once in the `shard.deps` loop. I believe
`register_link` is idempotent, in which case this will be fine (and CQ
will confirm that).

After ~2 weeks have passed and we're no longer testing release builds
from before this change, it'll be safe to remove the `shard.tests` for
loop from the recipes.

Bug: 51896
Change-Id: I4fa7e8956283beca37c95fbaa0fd7c6e45acc9d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389257
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>