[update-package] support packages.json version as int or string
This change modifies the update-package crate such that packages.json
files can have a version that's either a string or an int. I kept in
some of the "old" code so that it's easier to revert this in fxbug.dev/50754.

Fixed: 50713
Test: fx test -o update-package-tests
Change-Id: Iab96f385ed60aff97a8ae65a911e8a174e146c83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389348
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>
Testability-Review: Kevin Wells <kevinwells@google.com>