[modular] Improve logging around terminating session agents.
1) Log when we signal an agent to perform graceful termination.
2) Log when an agent has terminated expectedly during shutdonw.
3) Change the language around unexpected terminations & remove the
zx_status_t, as it was the status of the AgentController channel, which
is generated by Modular, not by the agent, and is misleading.

For (3), we should revisit this in the future to capture the
OnTerminated status from the ComponentController.

Test: existing
Change-Id: I5f34189d7ec02c5e301289c1e715066e2854f8db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388119
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Paweł Furtek <furtek@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Gabe Schine <thatguy@google.com>