[rust_crate] Move rouille from target to target + host.
Rouille provides a simple web framework that is also applicable to host
tools. This framework is planned to be used in an upcoming host tool so
moving it so it can be used both on the target and host.

Bug: 51992
Change-Id: Ic7a4d5b1222dd8c96d14eadddcb01e985b71dcc9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389393
Reviewed-by: Adam Perry <adamperry@google.com>
Testability-Review: Adam Perry <adamperry@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>