[nmh] Remove reachability from the core product
This reverts commit 86ffc3b21252637068770d8f1508289d2a82c31e.

Saves 344064 bytes from the prod build image size.

The thinking is that we're not using the primary functionality of the
component right now (reachability in Inspect output) and it can be
re-enabled once we've made enough room for NetworkManager and DHCPv6.

Blobstats for a prod (_eng) build before/after this:

Before:

$ fx blobstats | tail -n 3
   53% 178274304 / 381648896
$ fx blobstats | grep reachability
     Size      Prop   Private Name
   901120    341146    335872 reachability

After:

$ fx blobstats | tail -n 3
   53% 177930240 / 380739584
$ fx blobstats | grep reachability

Fixed: 52174
Change-Id: I818b5c570b599131d8b18e214b5a3610b7a23ee9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388864
Commit-Queue: Nicolas Sylvain <nsylvain@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>