[fdio] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow LocalVNode::Entry objects to exist in multiple WAVL trees
simultaneously.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Change-Id: I2369f2229a0d16f22df74f12d3b957ac196b5a10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386819
Commit-Queue: Nicolas Sylvain <nsylvain@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Testability-Review: Christopher Anderson <cja@google.com>