[carnelian] Demonstrate rotating app content
Modified the button sample to draw with one of
four rotations as an example of how the recovery
UI might accomplish a similar task.

Used this rotation requirement as a chance to
experiment with Euclid's typed geometry. So far
it seems to be a worthwhile feature, since it
helps prevent the mixing of display aligned and
presentation aligned coordinates.

While this change will allow recovery to present
its UI correctly, more design and implementation
will be needed to make apps in general function
correctly when running on the frame buffer on a
device with a rotated screen.

Testing: tested the button sample on the frame
buffer as well as under scenic. Ran all unit
tests.

Change-Id: I8e444d6e7956edeaad4f7632e62198d061ebb303
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383480
Commit-Queue: Rob Tsuk <robtsuk@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Testability-Review: David Reveman <reveman@google.com>