[netstack] Log DHCPv6 observations to Cobalt
...logging the events observed since the last time we logged to
Cobalt, while keeping polling logic outside the Cobalt client.

main gives the ndpDispatcher a `hasEvents` function to call, which
registers novelty with the Cobalt client.

When the Cobalt client is ready to log, it collects events from the
producer, allowing the caller to ensure that all novelty since the last
log is included without coupling the Cobalt client to the specific
business logic for that metric.

This ensures that we only log when we have to, and we only log what we
have to.

Fixed: 48858
Tested: fx test netstack_test -- -test.run ObservationsFromDHCPv6
Change-Id: I5d9962c38891dcb59196dca3c29ae1f16d187d2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374937
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>
Commit-Queue: Bert Muthalaly <stijlist@google.com>