[blobfs] Offload read-only blob connections to a zx::stream
After this CL, read-only connections to a blob are offloaded to a
zx::stream.  The zx::stream operations are still performed in the blobfs
process by libfs.  A future CL will offload these operations to the
client to improve performance.

This CL is a rebase/reland of
https://fuchsia-review.googlesource.com/c/fuchsia/+/369660

Bug: 45691
Test: No behavior change.
Change-Id: I518e82b3b26412c50f478c92b319a748c16bd464
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375441
Commit-Queue: Adam Barth <abarth@google.com>
Testability-Review: Adam Barth <abarth@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>