[intl] (Finally) implement the l10n message lookup
The implementation is no longer a dummy one: it will return
the translated message given the correct message ID.

The implementation is simplistic in that it assumes we know exactly
the list of languages we want to support.  This should be enough
for now.  Locale fallback is currently not done programmatically;
rather, we will package every locale and use pkgfs deduping to
remove ambiguities.

Fixed: 50725
Change-Id: I3dd23363c51fbd57dc253ecefdaeddfc7cfa3984
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388118
Testability-Review: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>