[fbl][sll] Reorder SLL template params
Reorder the existing SinglyLinkedList template parameters to put the
node traits at the back of the list.

There are a few reasons for doing this.  Primarily, there are not a
lot of good reasons any more to need to specify custom node traits
anymore (although, there are a few).  Putting this parameter in the
middle of the list of other parameters which user might actually want
to overload really serves no good purpose.

Second, this lets us allow the default node traits to know about the
TagType used to allow individuals to put objects into multiple
containers simultaniously.  This, in turn, allows us to eliminate an
awkward indirection between the node traits and the rest of the list
which was needed in order to allow the |node_state| method to accept a
template parameter, when user implementations of node_state had no
template paramters.  TL;DR, we don't need AddGenericNodeTraits<>
internally anymore.

Finally, this starts to align the SLL with the direction that the DLL
is headed in order to allow direct removal of nodes from lists.  While
the SLL will never allow this by nature of the data structure, it is
still good to have the patterns be as common as possible across the
types of containers.

Also, start to prefer the use of |PtrType| instead of |T| in the names
of the template parameters used here as it can be confusing.  T
in a context like this usually implies a specific object type, not a
specific type of pointer to an object that I want to store in a
container.  So, change the name now as part of the cleanup effort.

Tests: existing and updated unit tests
Change-Id: Id4ed8bd69401cac5554346132adb12a6998bd1b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386827
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>