[debugger] Support breakpoints conditional on hit-count
Conditional breakpoint is a breakpoint that only stops the execution
when some condition is met. It's useful in a loop where a breakpoint
could be hit thousands of times. This patch adds support for "hit-mult",
which means the breakpoint will only take effect when the hit-count is a
multiple of the given number.

An option for "break" command is added, together with a setting called
"hit-mult" which can be get/set on breakpoints.

Bug: 43929
Change-Id: Ib7d0e000da7d8c23c55c2368c0c4ce5eb9d0a6a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389614
Commit-Queue: Dangyi Liu <dangyi@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Testability-Review: Brett Wilson <brettw@google.com>