[ffx][overnet] Cancels overnet pipe if host-pipe command fails.
This adds additional tracking on the host pipe process to ensure that it
is cleanly torn down (including the reimplementation of std::io::copy
running in the background to helpn it communicate over FIDL channels).

This change came about as a result of debugging hanging issues that
originally seemed to be part of the daemon itself.

Change-Id: If35dbdacdf5945122aa8266058f1e57542de1ce5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386936
Testability-Review: Andrew Davies <awdavies@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>