[blobfs] Avoid inflating during compression and reduce threshold
This teaches the target and host side compression logic for blobs to
never write a compressed blob if it would be larger than the original.
We store a bit for each blob indicating whether it is compressed so
there's no reason to inflate a blob if the compression algorithm would
otherwise want to. The zstd algorithm will return larger outputs for
some small inputs in the configuration we are currently using.

This also reduces the minimum size for considering compression on a
blob to a single block, as we do benefit from compression for some
small files.

Test:
- Host compression behavior tested in blobfs-host-test
- Unit tests for small compressable and incompressable blobs in blobfs-unit-test

Bug: 39667
Co-authored-by: James Robinson <jamesr@google.com>
Change-Id: Ib8af41bd37a75b8bae648275516d4553e6e125a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389473
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Testability-Review: James Robinson <jamesr@google.com>