[fidl][llcpp] Avoid blocking on unbind.
The FIDL client takes a strong reference to the binding whenever it
needs to make an outgoing call on the channel. Currently, on deletion,
some thread will wait for all references to be released. However, as
clients may invoke zx_channel_call(), this wait can be unbounded.

This change makes it such that rather than waiting for a signal from the
destructor in some thread, the destructor will instead post a task which
will result in the user-specified unbound hook being executed.

As was previously the case, the user must ensure that the dispatcher is
valid through to executing the unbound hook. However, that is "actually"
required in all unbind paths, since they will all result in an
additional task being posted to the dispatcher to execute the hook.
Related to this, the unused AsyncBind() variant which takes a
std::unique_ptr<Interface> has been removed for the time being, as it
currently implicitly creates an unbound hook, which can cause issues if
the user chose to destroy the loop.

Bug: 48671
Test: runtests -t fidl-async-test-test -r 100 && runtests -t fidl-test
-r 10

Change-Id: Iaf420c3fc2e8fba2226bacfc57de5bb19a2c53f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388194
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Madhav Iyengar <madhaviyengar@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Testability-Review: Andres Oportus <andresoportus@google.com>