[packet] Clarify records parsing result values
The previous return type when parsing a record was confusing and not
intuitive. A new `ParsedRecord` type is introduced to clarify the result
of a record parsing operation, and the next action that should be taken.

No behaviour change.

Test: Existing tests pass.

Change-Id: I69a323ff2d883760d563b7e65664b78503cecfd3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390369
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Testability-Review: Joshua Liebow-Feeser <joshlf@google.com>