[size_checker] Add separate budget for shared libraries we distribute
The Fuchsia project builds and distributes a number of shared libraries
to partners through the SDK.  We also change the contents of these
libraries periodically.  Since they're coming from us, it makes more
sense to budget the size for these separately from the package or
packages that may use them. This allows a size budget specification to
define the set of file paths in packages that should be considered in
this category and an overall size limit for these based on their
expected size and propagation times.  If a dedicated size limit for
this category is not present in the input specification, this check is
ignored.

Change-Id: I0991db9041a044b4dd1b6288ccb055eda8837759
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390640
Reviewed-by: Gordon Tsai <gtsai@google.com>
Testability-Review: James Robinson <jamesr@google.com>
Commit-Queue: James Robinson <jamesr@google.com>