[netstack][test] Disable retransmit timeout test.
This test is waiting for the connection to timeout after multiple
retransmissions. After a recent change in upstream gvisor, this
timeout increases to ~20mins which is unreasonably long test time.

The TCP connection times out when there are N retransmits (15).
The retransmit intervals exponentially increase with a lower bound
of 1sec (or 200msec if there is a data-ack) and upper bound of 120sec.

We can enable this test when we have a way to tune the N value above,
thus reducing the test duration.
Similar to Linux sysctl net.ipv4.tcp_retries2

Bug: 52278

Change-Id: I847114c78eab9457dbff352ff99b1d504e3d6788
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390534
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>
Commit-Queue: Mithun Iyer <iyerm@google.com>