[power] Components can state why the power state is changing
This change is part of a larger effort to expose more granular
information on why device's reboot gracefully. It adds a method to the
Admin protocol that allows callers to specify why they are requesting
a change in the device's power state. Additionally, this change adds a
protocol that allows components to watch for requests to change the
system's power state by registering a callback with the server of
Admin.

Bug: 52190

Change-Id: Ic958d92c7f275536ae2434f880b465168824ce06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390219
Commit-Queue: Alex Pankhurst <pankhurst@google.com>
Reviewed-by: Patrick Shickel <pshickel@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Ruchira Ravoori <ravoorir@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
API-Review: Ruchira Ravoori <ravoorir@google.com>
Testability-Review: Francois Rousseau <frousseau@google.com>