[bt][avrcp-target] Return 0 if reference time is in the future.
It is valid for a client to provide a TimelineFunction in which the
reference_time is in the future. However, when calculating the
song_position, this can sometimes lead to negative song positions. Since
AVRCP only relays non-negative song positions, clamp the calculation to
zero.

Removed the warning that was issued if the reference_time was in the
future (i.e the diff was negative).

Fixed: 52318
Bug: 52254
Test: Updated
bt-avrcp-target-unittests: test_timeline_fn_to_song_position
Change-Id: I158d00211cb053bc32c17a9fa40a409d8612b7ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390759
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>