[async] Move async-default to its own folder.
This will allow it to get migrated to the GN build while the rest of async
is currently blocked by FIDL dependencies.

This change shedded some light on a few places in the codebase where the library
was not properly depended on.

Bug: 36548
Change-Id: I189ec2bc67971ce5a1880d8e19e776457a2d4667
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390786
Commit-Queue: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Testability-Review: Shai Barack <shayba@google.com>