[wlan][policy] Remove network
Adds the 'remove a saved network' functionality for the policy client
API. The credential and other data about the network will be forgotten,
so in order to auto-connect or manually connect through the client
provider API, it will need to be saved again.

Removing a network will cause a disconnect if we are currently connected
to the network through the new API. If we connected to the network
through the legacy code (including current auto-connection behavior),
removing the network will not trigger a disconnect.

Tests: added/modified unit tests
	 fx test wlancfg-tests
       manually tested remove using donut CLI tool and save/get,
       including disconnect when removed and persistence
Bug: 49581
Change-Id: Ide4d84002eb41d2f901efc364e4971579ce098a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378513
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Testability-Review: Jeff Belgum <belgum@google.com>