[isolated-ota] Support using omaha to install updates
In some instances we want to use Omaha to determine the update package
to install, along with the URL of the update package. When isolated-ota
is supplied an Omaha app id, make it use Omaha to get the update
metadata.

This CL adds a custom Omaha installer for isolated-ota which leverages
the existing sandboxed system-updater implementation to perform the
installation.

Change-Id: If2e6aa8cfdbda5b68fa0be95d83307699a8055b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386918
Testability-Review: Simon Shields <simonshields@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>