[kernel] Move ownership of root job to Executor.
Currently the RootJobObserver is responsible for creation of the root
job, and has the only reference to it. Methods in Executor pull out
the root job dispatcher on demand.

Move responsibility for ownership / creation of the root job to the
Executor class, passing it in to the RootJobObserver as an argument.

Bug: 51463
Change-Id: I90e6a0217ce8815ff272e8b1b404797180865430
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390974
Commit-Queue: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>