[aml][cpu] Add aml CPU as a child of the composite device
AML CPU wasn't being added as a child of the composite device
which was causing the driver manager to complain about a failure to
add a device in bind.

Tested by observing that the "failed to add device in bind" message
no longer appears in the logs and that the cpuctl functionality still
works.

Change-Id: Id53c4d5be8da9a008026857ce4270c28df630729
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382913
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>