[blobfs] Make compression level configurable
Adds --compression-level flag, which can be used to configure the
compression level blobfs uses for newly written blobs.

The flag is only respected for ZSTD* variants. A runtime error is raised
if the flag is set to an invalid value, or if it is set when a different
compression level is used.

Bug: 52261
Tested: Booted; Unit tests.
Change-Id: Ie625c8c5e610337de91e93eb15ac884bf974314a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391364
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>