[debugger] Handle line 0 stepping better.
A previous change to avoid reporting "line 0" entries in stack frames
caused stepping to stop at these entries. This change clarifies this
behavior in comments in ModuleSymbols, and fixes the behavior.

The solution is to remove the check that the stack frame's file/line
match the line entry for the given address. This was added to try to
detect the case where the entry was at the beginning of an inline frame.
but we can check this condition directly instead, and use the
information from the line table consistently.

Change-Id: I35ace5e608e211526e94799a08d2be800a321682
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390776
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>