[fidl][llcpp] Fix Close() path in Reply().
In AsyncTransaction::Reply(), errors cause the binding to be Close()d.
This can occur at a time when the AsyncTransaction has ownership of the
binding's internal self-reference. Previously, that self-reference was
not being returned to the binding prior to the Close(), potentially
leading to crashes. This CL fixes this behavior to restore the reference
and ensure that no invariants are broken.

Bug: 50822,42086
Test: runtests -t fidl-async-test-test -r 100
Change-Id: Ib8f798e53e670bf3c53d3a2d164ec536195656fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390639
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Andres Oportus <andresoportus@google.com>
Commit-Queue: Madhav Iyengar <madhaviyengar@google.com>