[media][audio] Fix flake in audio-loopback-tests
The audio-loopback-tests timed out when searching for the expected
packet in the loopback capture buffer. This is due to the packet data
being offset in the capture buffer beyond the expected payload offset.

To address this issue, the test now searches the capture buffer for the
true offset of the packet data, awaits all frames to be captured, and
then checks the capture buffer for the expected data, starting from the
true offset.

Multiply: audio-loopback-tests

Bug: 47968

Test: fx test audio-loopback-tests

Change-Id: I536c62ea56f951fcae3f57dfe0e175c5bfce5821
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/377074
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Payton Bradsky <pbradsky@google.com>
Testability-Review: Martin Puryear <mpuryear@google.com>