[fshost] Remove dependency on hermetic compute
This CL is part of a series of CLs that migrate the system away from the
hermetic compute library. Currently, the cost of maintaining hermetic
compute outweighs the benefits. Rather than decompressing the ZBI
out-of-process, we now decompress the ZBI in-process in fshost.

Change-Id: I2b2d5593aa619da85000450359da91d59fd4c95e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391419
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Testability-Review: Adam Barth <abarth@google.com>
Commit-Queue: Adam Barth <abarth@google.com>