[gidl] Generate multiple output files in GIDL for C++ targets
Currently, GIDL is generating a C++ file that takes a long time to
compile. This change splits up the file.

Originally, I hoped to name each output file based on the type name and
generate a GN dependency on that file. Unfortunately, this is quite
difficult to do in GN. For expediency, GIDL now generates a fixed number
of outputs, leaving unused output as empty files. This means specifying
the fixed number of outputs to be a larger number than the actual
generated outputs. However, if this number is exceeded, it is easy to
add additional outputs.

Bug: 52131

Change-Id: I4458e96d08f4189857cb91ae4caf27a921f24b89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390528
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>