[kernle][vm] Unify VMO Commit and Pin operations
Performing Pin as a separate operation on VMOs introduces the chance
for race conditions to cause the zx_bit_pin syscall to fail due to a
decommit happening between the Commit and the Pin.

Although Pin is allowed to have transient failures, doing so
unnecessarily places burden on clients and is less efficient overall.
In this case, preventing a race between Commit and Pin is a minimal
increase in complexity for the kernel, whilst greatly increasing the
Pin success rate to clients.

The included test case is not perfectly realistic since if clients
are Decommitting their own VMOs then they probably shouldn't epxect
to also be able to pin it. Howver, the test simulates the idea of
the kernel acting as the decommitter due to eviction, zero page
deduplication etc.

Change-Id: I781c336c46a1f96fb23c136455c0ab79f7ab3ba9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388093
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>